take relevant advice, changed some one
